ecommerce-live-coding-teamrando/requirements.txt
asgiref==3.6.0
black==23.1.0
click==8.1.3
Django==4.1.6
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
Faker==17.0.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
